+++ crates/ens/Cargo.toml
@@ -23,9 +23,13 @@ rustdoc-args = [
 workspace = true
 
 [dependencies]
-alloy-contract.workspace = true
+alloy-contract = { workspace = true, optional = true }
 alloy-primitives.workspace = true
-alloy-provider.workspace = true
-alloy-sol-types.workspace = true
-async-trait.workspace = true
-thiserror.workspace = true
+alloy-provider = { workspace = true, optional = true }
+alloy-sol-types = { workspace = true, optional = true }
+async-trait = { workspace = true, optional = true }
+thiserror = { workspace = true, optional = true }
+
+[features]
+contract = ["dep:alloy-contract", "dep:thiserror", "dep:alloy-sol-types"]
+provider = ["contract", "dep:alloy-provider", "dep:async-trait"]
